try to resolve conflits

+++ include/Triangle.h
@@ -12,21 +12,21 @@ class Triangle
 		Point get_p1();
 		Point get_p2();
 		Point get_p3();
-		Vector get_normal();
-		
+        Vector get_normal();
+
 		void set_p1(Point p1);
 		void set_p2(Point p2);
 		void set_p3(Point p3);
 		
-		Vector compute_normal();
-		double compute_D();
-		double compute_t(Vector source,Ray r);
+        Vector compute_normal();
+        float compute_D();
+        float compute_t(Vector source,Ray r);
         bool ray_intersect_plan(Ray r,Point& P);
         bool ray_intersect(Ray r,Point& P);
-        Point get_min_bounding_box();
+		Point get_min_bounding_box();
 		Point get_max_bounding_box();
 
 	private:
-		Point _p1, _p2, _p3;
-		Vector _normal;
+        Point _p1, _p2, _p3;
+        Vector _normal;
 };
